fastapi/requirements.txt
PyJWT
pandas
openai
pydantic
tiktoken
python-dotenv
fastapi[standard]
python-multipart
snowflake-connector-python